dom/xslt/tests/mochitest/test_bug1135764.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1135764
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1135764</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 1135764 **/
  SimpleTest.waitForExplicitFinish();
  var counter = 0;
  var startTimelineValue;

  function waitATick() {
    ++counter;
    if (counter == 1) {
      frames[0].requestAnimationFrame(waitATick);
      return;
    }
    ok(frames[0].document.timeline.currentTime !== startTimelineValue,
       "The timeline in an XSLT-transformed document should still advance");
    SimpleTest.finish();
  }
  addLoadEvent(function() {
    SpecialPowers.pushPrefEnv(
      {
        set: [
          ["dom.animations-api.timelines.enabled", true],
        ],
      },
      function() {
        var ifr = document.querySelector("iframe");
        ifr.onload = function() {
          startTimelineValue = frames[0].document.timeline.currentTime;
          frames[0].requestAnimationFrame(waitATick);
        };
        ifr.src = "file_bug1135764.xml";
      }
    );
  });
  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1135764">Mozilla Bug 1135764</a>
<p id="display">
  <iframe></iframe>
</p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
